+++ src/messages.js
@@ -2,24 +2,19 @@ import { debug } from "./util.js";
 
 class BaseMessage {
   constructor(actor, targetActor) {
-    /** @type {Actor5e} */
-    this.actor = actor;
-    /** @type {EffectChangeData[]} */
-    this.changes = this._getActiveEffectKeys(actor);
-    /** @type {EffectChangeData[]} */
-    this.targetChanges = this._getActiveEffectKeys(targetActor);
+    /** @type {object} */
+    this.actorFlags = this._getFlags(actor);
+    /** @type {object} */
+    this.targetFlags = this._getFlags(targetActor);
   }
 
-  _getActiveEffectKeys(actor) {
-    return actor
-      ? actor.appliedEffects
-          .flatMap((effect) => effect.changes)
-          .sort((a, b) => a.priority - b.priority)
-      : [];
+  _getFlags(actor) {
+    const flags = actor?.flags["adv-reminder"] || {};
+    return foundry.utils.flattenObject(flags);
   }
 
   get messageKeys() {
-    return ["flags.adv-reminder.message.all"];
+    return ["message.all"];
   }
 
   get targetKeys() {
@@ -34,24 +29,23 @@ class BaseMessage {
 
     // get messages from the actor and merge
     const keys = this.messageKeys;
-    const actorMessages = this.changes
-      .filter((change) => keys.includes(change.key))
-      .map((change) => change.value);
+    const actorMessages = Object.entries(this.actorFlags)
+      .filter(([key, messages]) => keys.includes(key))
+      .flatMap(([key, messages]) => messages);
     messages.push(...actorMessages);
 
     // get messages from the target and merge
     const targetKeys = this.targetKeys;
     if (targetKeys) {
-      const targetMessages = this.targetChanges
-        .filter((change) => targetKeys.includes(change.key))
-        .map((change) => change.value);
+      const targetMessages = Object.entries(this.targetFlags)
+        .filter(([key, messages]) => targetKeys.includes(key))
+        .flatMap(([key, messages]) => messages);
       messages.push(...targetMessages);
     }
 
     if (messages.length > 0) {
       debug("messages found:", messages);
       foundry.utils.setProperty(options, "options.adv-reminder.messages", messages);
-      foundry.utils.setProperty(options, "options.adv-reminder.rollData", this.actor.getRollData());
     }
   }
 }
@@ -69,18 +63,18 @@ export class AttackMessage extends BaseMessage {
   /** @override */
   get messageKeys() {
     return super.messageKeys.concat(
-      "flags.adv-reminder.message.attack.all",
-      `flags.adv-reminder.message.attack.${this.actionType}`,
-      `flags.adv-reminder.message.attack.${this.abilityId}`
+      "message.attack.all",
+      `message.attack.${this.actionType}`,
+      `message.attack.${this.abilityId}`
     );
   }
 
   /** @override */
   get targetKeys() {
     return [
-      "flags.adv-reminder.grants.message.attack.all",
-      `flags.adv-reminder.grants.message.attack.${this.actionType}`,
-      `flags.adv-reminder.grants.message.attack.${this.abilityId}`,
+      "grants.message.attack.all",
+      `grants.message.attack.${this.actionType}`,
+      `grants.message.attack.${this.abilityId}`,
     ];
   }
 }
@@ -95,7 +89,7 @@ class AbilityBaseMessage extends BaseMessage {
 
   /** @override */
   get messageKeys() {
-    return super.messageKeys.concat("flags.adv-reminder.message.ability.all");
+    return super.messageKeys.concat("message.ability.all");
   }
 }
 
@@ -103,8 +97,8 @@ export class AbilityCheckMessage extends AbilityBaseMessage {
   /** @override */
   get messageKeys() {
     return super.messageKeys.concat(
-      "flags.adv-reminder.message.ability.check.all",
-      `flags.adv-reminder.message.ability.check.${this.abilityId}`
+      "message.ability.check.all",
+      `message.ability.check.${this.abilityId}`
     );
   }
 }
@@ -113,8 +107,8 @@ export class AbilitySaveMessage extends AbilityBaseMessage {
   /** @override */
   get messageKeys() {
     return super.messageKeys.concat(
-      "flags.adv-reminder.message.ability.save.all",
-      `flags.adv-reminder.message.ability.save.${this.abilityId}`
+      "message.ability.save.all",
+      `message.ability.save.${this.abilityId}`
     );
   }
 }
@@ -122,7 +116,7 @@ export class AbilitySaveMessage extends AbilityBaseMessage {
 export class ConcentrationMessage extends AbilitySaveMessage {
   /** @override */
   get messageKeys() {
-    return super.messageKeys.concat("flags.adv-reminder.message.ability.concentration");
+    return super.messageKeys.concat("message.ability.concentration");
   }
 }
 
@@ -137,8 +131,8 @@ export class SkillMessage extends AbilityCheckMessage {
   /** @override */
   get messageKeys() {
     return super.messageKeys.concat(
-      "flags.adv-reminder.message.skill.all",
-      `flags.adv-reminder.message.skill.${this.skillId}`
+      "message.skill.all",
+      `message.skill.${this.skillId}`
     );
   }
 }
@@ -146,7 +140,7 @@ export class SkillMessage extends AbilityCheckMessage {
 export class InitiativeMessage extends AbilityBaseMessage {
   /** @override */
   get messageKeys() {
-    return super.messageKeys.concat("flags.adv-reminder.message.initiative");
+    return super.messageKeys.concat("message.initiative");
   }
 }
 
@@ -158,8 +152,8 @@ export class DeathSaveMessage extends AbilityBaseMessage {
   /** @override */
   get messageKeys() {
     return super.messageKeys.concat(
-      "flags.adv-reminder.message.ability.save.all",
-      "flags.adv-reminder.message.deathSave"
+      "message.ability.save.all",
+      "message.deathSave"
     );
   }
 }
@@ -175,16 +169,16 @@ export class DamageMessage extends BaseMessage {
   /** @override */
   get messageKeys() {
     return super.messageKeys.concat(
-      "flags.adv-reminder.message.damage.all",
-      `flags.adv-reminder.message.damage.${this.actionType}`
+      "message.damage.all",
+      `message.damage.${this.actionType}`
     );
   }
 
   /** @override */
   get targetKeys() {
     return [
-      "flags.adv-reminder.grants.message.damage.all",
-      `flags.adv-reminder.grants.message.damage.${this.actionType}`,
+      "grants.message.damage.all",
+      `grants.message.damage.${this.actionType}`,
     ];
   }
 }