+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 4.3.3
+
+- feature: set verified system version to 4.3.9
+- feature: update fast-forward check based on the system's implementation in `applyKeybindings`
+- feature: improve DAE integration with an AR Messages tab in the field browser and forcing all message flags to use Custom change mode
+- feature: refactor messages to use flags on the actor instead of reading the active effects
+- feature: refactor messages to use rollData from the roll instead of saving a copy itself
+
 # 4.3.2
 
 - feature: update Midi QOL compatibility since the big activity change, tested with version 12.4.64